# openapi.json
{
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "description": "多城市 × 单指标 预测对比",
        "examples": [
          {
            "city_names": [
              "深圳",
              "上海",
              "北京"
            ],
            "forecast_years": 5,
            "indicator": "gdp"
          }
        ],
        "properties": {
          "city_names": {
            "description": "城市名称列表",
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "City Names",
            "type": "array"
          },
          "forecast_years": {
            "default": 5,
            "description": "预测年数",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Forecast Years",
            "type": "integer"
          },
          "indicator": {
            "default": "gdp",
            "description": "要预测的指标,默认 gdp",
            "title": "Indicator",
            "type": "string"
          }
        },
        "required": [
          "city_names"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvinceRequest": {
        "description": "省级聚合 + 预测",
        "examples": [
          {
            "forecast_years": 5,
            "indicator": "gdp"
          }
        ],
        "properties": {
          "forecast_years": {
            "default": 5,
            "description": "预测年数",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Forecast Years",
            "type": "integer"
          },
          "indicator": {
            "default": "gdp",
            "description": "要预测的指标",
            "title": "Indicator",
            "type": "string"
          }
        },
        "title": "ProvinceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "中国城市经济数据 REST API：35 城、16 年指标、集成预测与企业/政府/产业分析。",
    "license": {
      "name": "GPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "title": "Urban Pulse",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/forecast/compare": {
      "post": {
        "description": "对比多个城市 × 同一指标 的预测结果。",
        "operationId": "compare_city_forecasts_api_v1_forecast_compare_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare City Forecasts Api V1 Forecast Compare Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "多城市 × 单指标 预测对比",
        "tags": [
          "预测"
        ]
      }
    },
    "/api/v1/forecast/gdp/{city_name}": {
      "get": {
        "description": "兼容旧端点:预测指定城市 GDP。\n内部走新的 forecast_city_indicator 逻辑,真实置信区间。",
        "operationId": "forecast_city_gdp_api_v1_forecast_gdp__city_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_name",
            "required": true,
            "schema": {
              "title": "City Name",
              "type": "string"
            }
          },
          {
            "description": "预测年数",
            "in": "query",
            "name": "forecast_years",
            "required": false,
            "schema": {
              "default": 5,
              "description": "预测年数",
              "maximum": 20,
              "minimum": 1,
              "title": "Forecast Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast City Gdp Api V1 Forecast Gdp  City Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "[兼容] 预测城市 GDP",
        "tags": [
          "预测"
        ]
      }
    },
    "/api/v1/forecast/indicator/{city_name}": {
      "get": {
        "description": "预测指定城市 × 任意指标的未来 years 年值。",
        "operationId": "forecast_city_indicator_endpoint_api_v1_forecast_indicator__city_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_name",
            "required": true,
            "schema": {
              "title": "City Name",
              "type": "string"
            }
          },
          {
            "description": "指标代码,支持: data_quality, fiscal_revenue, gdp, gdp_growth, industry_high_tech_ratio, land_price, local_support_rate, policy_coverage, population, rd_intensity, rd_subsidy, supplier_count, tax_reduction",
            "in": "query",
            "name": "indicator",
            "required": false,
            "schema": {
              "default": "gdp",
              "description": "指标代码,支持: data_quality, fiscal_revenue, gdp, gdp_growth, industry_high_tech_ratio, land_price, local_support_rate, policy_coverage, population, rd_intensity, rd_subsidy, supplier_count, tax_reduction",
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "description": "预测年数",
            "in": "query",
            "name": "forecast_years",
            "required": false,
            "schema": {
              "default": 5,
              "description": "预测年数",
              "maximum": 20,
              "minimum": 1,
              "title": "Forecast Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast City Indicator Endpoint Api V1 Forecast Indicator  City Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "预测城市任意指标",
        "tags": [
          "预测"
        ]
      }
    },
...
    "/api/v1/forecast/province/all": {
      "get": {
        "description": "一次性对所有 7 个省份 × 1 个指标做\"过去发展找规律 + 未来预测\"。\n\n适合:\n- 数学建模课程的\"按省份 GDP 演化\"展示\n- 政策评估报告的横向对比\n- 政府决策者一页看完全国趋势\n\n注意:此路由必须在 `/province/{province_name}` 之前,否则会被路径参数\n`province_name=\"all\"` 抢先匹配。FastAPI 按声明顺序匹配,先到先得。",
        "operationId": "forecast_all_provinces_endpoint_api_v1_forecast_province_all_get",
        "parameters": [
          {
            "description": "要预测的指标",
            "in": "query",
            "name": "indicator",
            "required": false,
            "schema": {
              "default": "gdp",
              "description": "要预测的指标",
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "description": "预测年数",
            "in": "query",
            "name": "forecast_years",
            "required": false,
            "schema": {
              "default": 5,
              "description": "预测年数",
              "maximum": 20,
              "minimum": 1,
              "title": "Forecast Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast All Provinces Endpoint Api V1 Forecast Province All Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "全 7 省 × 单指标 批量预测报告",
        "tags": [
          "预测"
        ]
      }
    },
    "/api/v1/forecast/province/{province_name}": {
      "post": {
        "description": "对指定省份 × 指定指标,先把省内各城市历史数据按\"绝对量求和/率加权\"\n聚合,再对未来 years 年做预测。",
        "operationId": "forecast_province_endpoint_api_v1_forecast_province__province_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "province_name",
            "required": true,
            "schema": {
              "title": "Province Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvinceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast Province Endpoint Api V1 Forecast Province  Province Name  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "省级 × 单指标 聚合 + 预测",
        "tags": [
          "预测"
        ]
      }
    },
...
    "/api/v1/forecast/report/full": {
      "get": {
        "description": "城市级完整预测报告 — 8 个城市 × 单指标批量预测。",
        "operationId": "get_full_forecast_report_api_v1_forecast_report_full_get",
        "parameters": [
          {
            "description": "默认 gdp",
            "in": "query",
            "name": "indicator",
            "required": false,
            "schema": {
              "default": "gdp",
              "description": "默认 gdp",
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "forecast_years",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Forecast Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Full Forecast Report Api V1 Forecast Report Full Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "完整预测报告(8 城 × 1 指标)",
        "tags": [
          "预测"
        ]
      }
    },
...
    "/api/v1/regions/{code}/forecast/{indicator}": {
      "get": {
        "description": "对指定区域（目前仅 city 层级实现）的某指标做未来 N 年预测。",
        "operationId": "forecast_region_indicator_api_v1_regions__code__forecast__indicator__get",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "indicator",
            "required": true,
            "schema": {
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "description": "预测年数",
            "in": "query",
            "name": "forecast_years",
            "required": false,
            "schema": {
              "default": 5,
              "description": "预测年数",
              "maximum": 20,
              "minimum": 1,
              "title": "Forecast Years",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast Region Indicator Api V1 Regions  Code  Forecast  Indicator  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "预测区域指标",
        "tags": [
          "区域"
        ]
      }
    },
...
  },
...
}